+++ plugins/Winnow.Plugin.SteamGridDb/Package.ps1
@@ -1,5 +1,5 @@
 param(
-    [string]$OutputDirectory = (Join-Path $PSScriptRoot '..\..\artifacts\xbox-plugin')
+    [string]$OutputDirectory = (Join-Path $PSScriptRoot '../../artifacts/steamgriddb-plugin')
 )
 
 & (Join-Path $PSScriptRoot '../../packaging/New-PluginPackage.ps1') -PluginDirectory $PSScriptRoot -OutputDirectory $OutputDirectory